test(research): split PDF equation region coverage
